Get Cell Ontology

A thin wrapper around get_ontology to get the specific version of the Cell Ontology used in the original analyses that produced the results stored in load_example_results.

get_cl(name = "cl", tag = "v2023-09-21", lvl = 1, remove_rings = TRUE, ...)

Arguments

name

<...>

Any ontology name from get_ols_options

"mondo"

Import the Mondo ontology. GitHub.

.
"hpo"

Import the Human Phenotype Ontology. GitHub.

tag

tag for the GitHub release to which this data should be attached.

lvl

Depth of the ancestor terms to add. Will get the closest ancestor to this level if none have this exact level.

remove_rings

Parameter passed to simona::import_ontology.

...

Arguments passed on to KGExplorer::get_ontology

save_dir

Directory to save a file to.

force_new

If TRUE, force a new download.

method

Compute ontology levels using:

"height" (default)

dag_height.

"depth"

dag_depth.

filetype

File type to search for.

import_func

Function to import the ontology with. If NULL, automatically tries to choose the correct function.

terms

A vector of ontology term IDs.

add_metadata

Add metadata to the resulting ontology object.

add_n_edges

Add the number of edges (connections) for each term.

add_ontology_levels

Add the ontology level for each term.

Examples

cl <- get_cl()
#> Translating ontology terms to ids.
#> Using cached ontology file (1/1):
#>  /github/home/.cache/R/KGExplorer/ontologies/github/cl_v2023-09-21.rds
#> Keeping descendants of 2 term(s).
#> 2,711 terms remain after filtering.