get valid gene lists

Gene lists for EWCE must have 4 or more genes. These genes must also be present in the CTD file (see EWCE docs), and must be unique (occasionally a gene list may contain the same gene repeated. This would cause an error).

get_valid_gene_lists(
  ctd,
  list_names,
  list_name_column,
  gene_data,
  annotLevel = 1,
  min_genes = 4,
  verbose = TRUE
)

Arguments

ctd

CellTypeDataset generated using generate_celltype_data.

list_names

character vector of gene list names.

list_name_column

The name of the gene_data column that has the gene list names.

gene_data

data frame of gene list names and genes (see get_gene_lists).

annotLevel

An integer indicating which level of sct_data to analyse (Default: 1).

min_genes

Minimum number of genes per list (default: 4)

verbose

Print messages.

Value

A character vector of list_names that are associated with a valid number of genes

Details

This function gets gene lists that have at least four unique genes that are also present in the CTD file.