Load example results

This loads a example of enrichment results from gen_results using the Human Phenotype Ontology and a given CellTypeDataset (CTD). Results were then merged together with merge_results.

load_example_results(
  file = c("phenomix_results.tsv.gz"),
  multi_dataset = FALSE,
  tag = "latest",
  save_dir = KGExplorer::cache_dir(package = "MSTExplorer"),
  force_new = FALSE
)

Source

d <- "~/Desktop/ewce/rare_disease_celltyping_apps/cell_select" #### Descartes_All_Results_extras #### f0 <- file.path(d,"data/Descartes_All_Results_extras.rds") r0 <- readRDS(f0) data.table::setnames(r0,"list","hpo_name") r0$HPO_id=NULL r0 <- HPOExplorer:::fix_hpo_ids(dt=r0) f0new <- file.path(tempdir(),basename(f0)) saveRDS(r0,file = f0new) piggyback::pb_upload(file = f0new, tag = "v0.0.1", repo = "neurogenomics/MSTExplorer") #### Descartes_All_Results #### f1 <- file.path(d,"data/Descartes_All_Results.rds") r1 <- readRDS(f1) data.table::setnames(r1,"list","hpo_name") r1 <- HPOExplorer:::fix_hpo_ids(dt=r1) f1new <- file.path(tempdir(),basename(f1)) saveRDS(r1,file = f1new) piggyback::pb_upload(file = f1new, tag = "v0.0.1", repo = "neurogenomics/MSTExplorer") #### tabulamuris_merged #### f2 <- file.path(d,"data/tabulamuris_merged.rds") r2 <- readRDS(f2) data.table::setnames(r2,"list","hpo_name") r2 <- HPOExplorer:::fix_hpo_ids(dt=r2) f2new <- file.path(tempdir(),basename(f2)) saveRDS(r2,file = f2new) piggyback::pb_upload(file = f2new, tag = "v0.0.1", repo = "neurogenomics/MSTExplorer")

Results located in 'results' folder.

Arguments

file

File to load:

"Descartes_All_Results_extras.symptoms.full_join.rds"

Contains cell type-phenotype and cell type-symptom (phenotype + disease) enrichment results merged into one table. Used the Descartes CTD ( annotation level 1).

"Descartes_All_Results_extras.rds"

Contains cell type-phenotype enrichment results. Used the Descartes CTD ( annotation level 1).

"tabulamuris_merged.rds"

Contains cell type-phenotype enrichment results. Used the Tabula Muris CTD.

multi_dataset

Merge results generated from multiple CellTypeDataset references (e.g. DescartesHuman and HumanCellLandscape).

tag

tag for the GitHub release to which this data should be attached.

save_dir

Where to store the file locally.

force_new

Download the file even when a local copy already exists.

Value

dataframe of enrichment results.

Examples

res <- load_example_results()