Plot differential outcomes

Plot differential outcomes (annotations) of different diseases as they occur in different cell types via particular phenotypes.

plot_differential_outcomes(
  results,
  filters = NULL,
  facet_var = "disease_name",
  x_var = "celltype_symptom",
  y_var = "severity_score_gpt",
  remove_facets = NULL,
  run_stats = FALSE,
  max_facets = NULL,
  q_threshold = list(summary = 0.05, pairwise = NULL),
  ...
)

Arguments

results

The cell type-phenotype enrichment results generated by gen_results and merged together with merge_results

filters

A named list, where each element in the list is the name of a column in the data, and the vector within each element represents the values to include in the final data.

facet_var

Variable to facet by.

x_var

Variable to plot on the x-axis.

y_var

Variable to plot on the y-axis.

remove_facets

A character vector of facets to remove.

run_stats

If TRUE, run statistical tests and display the summary statistics directly on the plot.

max_facets

Maximum number of facets to display.

q_threshold

A list of thresholds for significance testing.

...

Arguments passed on to ggstatsplot::ggbetweenstats

data

A data frame (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted. Additionally, grouped data frames from {dplyr} should be ungrouped before they are entered as data.

x

The grouping (or independent) variable from data. In case of a repeated measures or within-subjects design, if subject.id argument is not available or not explicitly specified, the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted, the results can be inaccurate when there are more than two levels in x and there are NAs present. The data is expected to be sorted by user in subject-1, subject-2, ..., pattern.

y

The response (or outcome or dependent) variable from data.

type

A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

pairwise.display

Decides which pairwise comparisons to display. Available options are:

"significant" (abbreviation accepted: "s")

"non-significant" (abbreviation accepted: "ns")

"all"

You can use this argument to make sure that your plot is not uber-cluttered when you have multiple groups being compared and scores of pairwise comparisons being displayed. If set to "none", no pairwise comparisons will be displayed.

p.adjust.method

Adjustment method for p-values for multiple comparisons. Possible methods are: "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".

effsize.type

Type of effect size needed for parametric tests. The argument can be "eta" (partial eta-squared) or "omega" (partial omega-squared).

bf.prior

A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates. In addition to numeric arguments, several named values are also recognized: "medium", "wide", and "ultrawide", corresponding to r scale values of 1/2, sqrt(2)/2, and 1, respectively. In case of an ANOVA, this value corresponds to scale for fixed effects.

bf.message

Logical that decides whether to display Bayes Factor in favor of the null hypothesis. This argument is relevant only for parametric test (Default: TRUE).

results.subtitle

Decides whether the results of statistical tests are to be displayed as a subtitle (Default: TRUE). If set to FALSE, only the plot will be returned.

xlab

Label for x axis variable. If NULL (default), variable name for x will be used.

ylab

Labels for y axis variable. If NULL (default), variable name for y will be used.

caption

The text for the plot caption. This argument is relevant only if bf.message = FALSE.

title

The text for the plot title.

subtitle

The text for the plot subtitle. Will work only if results.subtitle = FALSE.

digits

Number of digits for rounding or significant figures. May also be "signif" to return significant figures or "scientific" to return scientific notation. Control the number of digits by adding the value as suffix, e.g. digits = "scientific4" to have scientific notation with 4 decimal places, or digits = "signif5" for 5 significant figures (see also signif()).

var.equal

a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

conf.level

Scalar between 0 and 1 (default: 95% confidence/credible intervals, 0.95). If NULL, no confidence intervals will be computed.

nboot

Number of bootstrap samples for computing confidence interval for the effect size (Default: 100L).

tr

Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

centrality.plotting

Logical that decides whether centrality tendency measure is to be displayed as a point with a label (Default: TRUE). Function decides which central tendency measure to show depending on the type argument.

mean for parametric statistics

median for non-parametric statistics

trimmed mean for robust statistics

MAP estimator for Bayesian statistics

If you want default centrality parameter, you can specify this using centrality.type argument.

centrality.type

Decides which centrality parameter is to be displayed. The default is to choose the same as type argument. You can specify this to be:

"parameteric" (for mean)

"nonparametric" (for median)

robust (for trimmed mean)

bayes (for MAP estimator)

Just as type argument, abbreviations are also accepted.

centrality.point.args,centrality.label.args

A list of additional aesthetic arguments to be passed to ggplot2::geom_point() and ggrepel::geom_label_repel() geoms, which are involved in mean plotting.

point.args

A list of additional aesthetic arguments to be passed to the ggplot2::geom_point().

boxplot.args

A list of additional aesthetic arguments passed on to ggplot2::geom_boxplot(). By default, the whiskers extend to 1.5 times the interquartile range (IQR) from the box (Tukey-style). To customize whisker length, you can use the coef parameter, e.g., boxplot.args = list(coef = 3) for whiskers extending to 3 * IQR, or boxplot.args = list(coef = 0) to show only the range of the data.

violin.args

A list of additional aesthetic arguments to be passed to the ggplot2::geom_violin().

ggsignif.args

A list of additional aesthetic arguments to be passed to ggsignif::geom_signif().

ggtheme

A {ggplot2} theme. Default value is theme_ggstatsplot(). Any of the {ggplot2} themes (e.g., ggplot2::theme_bw()), or themes from extension packages are allowed (e.g., ggthemes::theme_fivethirtyeight(), hrbrthemes::theme_ipsum_ps(), etc.). But note that sometimes these themes will remove some of the details that {ggstatsplot} plots typically contains. For example, if relevant, ggbetweenstats() shows details about multiple comparison test as a label on the secondary Y-axis. Some themes (e.g. ggthemes::theme_fivethirtyeight()) will remove the secondary Y-axis and thus the details as well.

package,palette

Name of the package from which the given palette is to be extracted. The available palettes and packages can be checked by running View(paletteer::palettes_d_names).

ggplot.component

A ggplot component to be added to the plot prepared by {ggstatsplot}. This argument is primarily helpful for grouped_ variants of all primary functions. Default is NULL. The argument should be entered as a {ggplot2} function or a list of {ggplot2} functions.

Value

R object.

Examples

results <- load_example_results()
## Reduce the CTD list for example purposes
ctd_list <- load_example_ctd("ctd_DescartesHuman.rds",
                             multi_dataset = TRUE)
#> Loading ctd_DescartesHuman.rds
results <- add_symptom_results(results=results[ctd=="DescartesHuman"],
                               ctd_list=ctd_list)
#> Adding symptom-level results.
#> Reading cached RDS file: phenotype_to_genes.txt
#> + Version: v2025-11-24
#> Subsetting results by q_threshold and effect.
#> 19,929 associations remain after filtering.
#> Adding genes and disease IDs.
#> Reading cached RDS file: phenotype_to_genes.txt
#> + Version: v2025-11-24
## Reduce the number of diseases for example purposes
results <- results[disease_id %in% unique(results$disease_id)[seq(6)]]
#### Multiple phenotypes per disease #####
results <- HPOExplorer::add_gpt_annotations(results)
#> Translating ontology terms to ids.
#> Reading cached RDS file: phenotype_to_genes.txt
#> + Version: v2025-11-24
#> 151 phenotypes do not have matching HPO IDs.
#> Reading in GPT annotations for 16,982 phenotypes.
p1 <- plot_differential_outcomes(results,
                           facet_var = "disease_name",
                           y_var = "severity_score_gpt")
#> Adding disease_name and disease_description.
#> Loading required namespace: tidytext