Validate associations with the Monarch Knowledge Graph

This function validates the associations between phenotypes and cell types with the Monarch Knowledge Graph (MKG). It computes the number of phenotypes and cell types that are present in the MKG and the number of associations that are present in the MKG.

validate_associations_mkg(
  results = load_example_results(),
  kg = get_data("monarch_kg_cells.csv"),
  q_threshold = 0.05,
  cl = get_cl(),
  alt_ids = c(`CL:0000111` = "CL:2000032"),
  metric = c("dist_nca.min_adj", "dist_lca.min_adj", "dist_nca.min", "dist_lca.min")[1]
)

Arguments

results

The cell type-phenotype enrichment results generated by gen_results and merged together with merge_results

kg

A data.table with phenotype-cell type relationships ("from" and "to" columns, respectively) gathered from the Monarch Knowledge Graph.

q_threshold

The q value threshold to subset the results by.

cl

Cell Ontology object.

alt_ids

Alternative IDs to map CL IDs into.

metric

Ontological distance metric to use for plotting.

Examples

results <- load_example_results()
kg <- validate_associations_mkg(result=results)
#> Adding logFC column.
#> ℹ All local files already up-to-date!
#> Remaining: 82 phenotypes across 65 celltypes.
#> 90.24% phenotypes recovered.
#> Mapping cell types to cell ontology terms.
#> Adding stage information.
#> [i=1] cl_id: CL:0000000 & hpo_id: HP:0025354
#> Translating ontology terms to ids.
#> Using cached ontology file (1/1):
#>  /github/home/.cache/R/KGExplorer/ontologies/github/cl_v2023-09-21.rds
#> Keeping descendants of 2 term(s).
#> 2,711 terms remain after filtering.
#> [i=2] cl_id: CL:0000000 & hpo_id: HP:0005374
#> [i=3] cl_id: CL:0000015 & hpo_id: HP:0012863
#> [i=4] cl_id: CL:0000019 & hpo_id: HP:0000027
#> [i=5] cl_id: CL:0000019 & hpo_id: HP:0012864
#> [i=6] cl_id: CL:0000038 & hpo_id: HP:0012131
#> [i=7] cl_id: CL:0000066 & hpo_id: HP:0032599
#> [i=8] cl_id: CL:0000084 & hpo_id: HP:0002843
#> [i=9] cl_id: CL:0000084 & hpo_id: HP:0011840
#> [i=10] cl_id: CL:0000094 & hpo_id: HP:0001911
#> [i=11] cl_id: CL:0000097 & hpo_id: HP:0100494
#> [i=12] cl_id: CL:0000100 & hpo_id: HP:0002450
#> [i=13] cl_id: CL:2000032 & hpo_id: HP:0003134
#> [i=14] cl_id: CL:0000125 & hpo_id: HP:0100705
#> [i=15] cl_id: CL:0000127 & hpo_id: HP:0100707
#> [i=16] cl_id: CL:0000128 & hpo_id: HP:0100706
#> [i=17] cl_id: CL:0000169 & hpo_id: HP:0006279
#> [i=18] cl_id: CL:0000169 & hpo_id: HP:0006274
#> [i=19] cl_id: CL:0000178 & hpo_id: HP:0010789
#> [i=20] cl_id: CL:0000182 & hpo_id: HP:0031137
#> [i=21] cl_id: CL:0000187 & hpo_id: HP:0100299
#> [i=22] cl_id: CL:0000188 & hpo_id: HP:0004303
#> [i=23] cl_id: CL:0000232 & hpo_id: HP:0001876
#> [i=24] cl_id: CL:0000232 & hpo_id: HP:0001877
#> [i=25] cl_id: CL:0000232 & hpo_id: HP:0410064
#> [i=26] cl_id: CL:0000232 & hpo_id: HP:0410063
#> [i=27] cl_id: CL:0000232 & hpo_id: HP:0020054
#> [i=28] cl_id: CL:0000232 & hpo_id: HP:0012187
#> [i=29] cl_id: CL:0000233 & hpo_id: HP:0001872
#> [i=30] cl_id: CL:0000235 & hpo_id: HP:0004311
#> [i=31] cl_id: CL:0000236 & hpo_id: HP:0005372
#> [i=32] cl_id: CL:0000236 & hpo_id: HP:0002846
#> [i=33] cl_id: CL:0000492 & hpo_id: HP:0500267
#> [i=34] cl_id: CL:0000540 & hpo_id: HP:0002493
#> [i=35] cl_id: CL:0000540 & hpo_id: HP:0002529
#> [i=36] cl_id: CL:0000540 & hpo_id: HP:0012757
#> [i=37] cl_id: CL:0000542 & hpo_id: HP:0004332
#> [i=38] cl_id: CL:0000547 & hpo_id: HP:0025035
#> [i=39] cl_id: CL:0000556 & hpo_id: HP:0012143
#> [i=40] cl_id: CL:0000576 & hpo_id: HP:0012144
#> [i=41] cl_id: CL:0000586 & hpo_id: HP:0012862
#> [i=42] cl_id: CL:0000623 & hpo_id: HP:0012177
#> [i=43] cl_id: CL:0000623 & hpo_id: HP:0012176
#> [i=44] cl_id: CL:0000624 & hpo_id: HP:0031392
#> [i=45] cl_id: CL:0000625 & hpo_id: HP:0031393
#> [i=46] cl_id: CL:0000738 & hpo_id: HP:0008318
#> [i=47] cl_id: CL:0000738 & hpo_id: HP:0004852
#> [i=48] cl_id: CL:0000738 & hpo_id: HP:0001881
#> [i=49] cl_id: CL:0000763 & hpo_id: HP:0020047
#> [i=50] cl_id: CL:0000764 & hpo_id: HP:0012130
#> [i=51] cl_id: CL:0000765 & hpo_id: HP:0034278
#> [i=52] cl_id: CL:0000766 & hpo_id: HP:0010974
#> [i=53] cl_id: CL:0000767 & hpo_id: HP:0001912
#> [i=54] cl_id: CL:0000771 & hpo_id: HP:0001879
#> [i=55] cl_id: CL:0000775 & hpo_id: HP:0001874
#> [i=56] cl_id: CL:0000787 & hpo_id: HP:0030373
#> [i=57] cl_id: CL:0000792 & hpo_id: HP:0020111
#> [i=58] cl_id: CL:0000798 & hpo_id: HP:0500270
#> [i=59] cl_id: CL:0000813 & hpo_id: HP:0032182
#> [i=60] cl_id: CL:0000818 & hpo_id: HP:0030379
#> [i=61] cl_id: CL:0000835 & hpo_id: HP:0004808
#> [i=62] cl_id: CL:0000845 & hpo_id: HP:0030383
#> [i=63] cl_id: CL:0000891 & hpo_id: HP:0003651
#> [i=64] cl_id: CL:0000895 & hpo_id: HP:0410373
#> [i=65] cl_id: CL:0000898 & hpo_id: HP:0031396
#> [i=66] cl_id: CL:0000909 & hpo_id: HP:0410380
#> [i=67] cl_id: CL:0000912 & hpo_id: HP:0500263
#> [i=68] cl_id: CL:0000972 & hpo_id: HP:0030386
#> [i=69] cl_id: CL:0001050 & hpo_id: HP:0410395
#> [i=70] cl_id: CL:0001062 & hpo_id: HP:0020177
#> [i=71] cl_id: CL:0002092 & hpo_id: HP:0005561
#> [i=72] cl_id: CL:0002191 & hpo_id: HP:0012135
#> [i=73] cl_id: CL:0002198 & hpo_id: HP:0005152
#> [i=74] cl_id: CL:0002319 & hpo_id: HP:0002282
#> [i=75] cl_id: CL:0002368 & hpo_id: HP:0012261
#> [i=76] cl_id: CL:0002422 & hpo_id: HP:0004312
#> [i=77] cl_id: CL:0002559 & hpo_id: HP:0500262
#> [i=78] cl_id: CL:0008002 & hpo_id: HP:0030089
#> [i=79] cl_id: CL:0008002 & hpo_id: HP:0003713
#> [i=80] cl_id: CL:0008002 & hpo_id: HP:0012084
#> [i=81] cl_id: CL:0011001 & hpo_id: HP:0006802
#> [i=82] cl_id: CL:0011025 & hpo_id: HP:0041080